+++ src/modules/perception/__init__.py
@@ -3,5 +3,6 @@ from .camera_capture import CameraCapture
 from .audio_capture import AudioCapture
 from .face_detector import FaceDetector, FaceData
 from .face_counter import FaceCounter
+from .gesture_controller import GestureController, GestureData
 
-__all__ = ['CameraCapture', 'AudioCapture', 'FaceDetector', 'FaceData', 'FaceCounter']
+__all__ = ['CameraCapture', 'AudioCapture', 'FaceDetector', 'FaceData', 'FaceCounter', 'GestureController', 'GestureData']